Made the test wait until the web server is handling connections rather than waiting a fixed amount of time. Switched from launching a process to running a thread.
